Chore: Refactoring datasets.py step 2 - fixations and scanpaths in own file
