+++ pysaliency/datasets/scanpaths.py
@@ -1,0 +1,177 @@
+import json
+from typing import Dict, List, Optional, Union
+
+import numpy as np
+from boltons.cacheutils import cached
+
+from ..utils.variable_length_array import VariableLengthArray, concatenate_variable_length_arrays
+from .utils import create_hdf5_dataset, decode_string, hdf5_wrapper, _load_attribute_dict_from_hdf5
+
+
+class Scanpaths(object):
+    """
+    Represents a collection of scanpaths.
+
+    Attributes:
+        xs (VariableLengthArray): The x-coordinates of the scanpaths.
+        ys (VariableLengthArray): The y-coordinates of the scanpaths.
+        n (np.ndarray): The image index
+        lengths (np.ndarray): The lengths of each scanpath.
+        scanpath_attributes (dict): Additional attributes associated with the scanpaths.
+        fixation_attributes (dict): Additional attributes associated with the fixations in the scanpaths.
+        attribute_mapping (dict): Mapping of attribute names to their corresponding values, will be used when creating `Fixations` instances from the `Scanpaths` instance.
+             for example {'durations': 'duration'}
+    """
+
+    xs: VariableLengthArray
+    ys: VariableLengthArray
+    n: np.ndarray
+
+    def __init__(self,
+                 xs: Union[np.ndarray, VariableLengthArray],
+                 ys: Union[np.ndarray, VariableLengthArray],
+                 n: np.ndarray,
+                 lengths=None,
+                 scanpath_attributes: Optional[Dict[str, np.ndarray]] = None,
+                 fixation_attributes: Optional[Dict[str, Union[np.ndarray, VariableLengthArray]]]=None,
+                 attribute_mapping=Dict[str, str]):
+
+        self.n = np.asarray(n)
+
+        if not isinstance(xs, VariableLengthArray):
+            self.xs = VariableLengthArray(xs, lengths)
+        else:
+            self.xs = xs
+
+        if lengths is not None:
+            if not np.all(self.xs.lengths == lengths):
+                raise ValueError("Lengths of xs and lengths do not match")
+
+        self.lengths = self.xs.lengths.copy()
+
+        self.ys = self._as_variable_length_array(ys)
+
+        if not len(self.xs) == len(self.ys) == len(self.n):
+            raise ValueError("Length of xs, ys, ts and n has to match")
+
+        # setting scanpath attributes
+
+        scanpath_attributes = scanpath_attributes or {}
+        self.scanpath_attributes = {key: np.array(value) for key, value in scanpath_attributes.items()}
+
+        for key, value in self.scanpath_attributes.items():
+            if not len(value) == len(self.xs):
+                raise ValueError(f"Length of scanpath attribute {key} has to match number of scanpaths, but got {len(value)} != {len(self.xs)}")
+
+        # setting fixation attributes
+
+        fixation_attributes = fixation_attributes or {}
+
+        self.fixation_attributes = {key: self._as_variable_length_array(value) for key, value in fixation_attributes.items()}
+
+        self.attribute_mapping = attribute_mapping or {}
+
+    def _check_lengths(self, other: VariableLengthArray):
+        if not len(self) == len(other):
+            raise ValueError("Length of scanpaths has to match")
+        if not np.all(self.lengths == other.lengths):
+            raise ValueError("Lengths of scanpaths have to match")
+
+    def _as_variable_length_array(self, data: Union[np.ndarray, VariableLengthArray]) -> VariableLengthArray:
+        if not isinstance(data, VariableLengthArray):
+            data = VariableLengthArray(data, self.lengths)
+
+        self._check_lengths(data)
+
+        return data
+
+    def __len__(self):
+        return len(self.xs)
+
+    @property
+    def ts(self) -> VariableLengthArray:
+        return self.fixation_attributes['ts']
+
+    @property
+    def subject(self) -> VariableLengthArray:
+        return self.scanpath_attributes['subject']
+
+
+    @hdf5_wrapper(mode='w')
+    def to_hdf5(self, target):
+        """ Write scanpaths to hdf5 file or hdf5 group
+        """
+        target.attrs['type'] = np.string_('Scanpaths')
+        target.attrs['version'] = np.string_('1.0')
+
+        target.create_dataset('xs', data=self.xs._data)
+        target.create_dataset('ys', data=self.ys._data)
+        target.create_dataset('n', data=self.n)
+        target.create_dataset('lengths', data=self.lengths)
+
+        scanpath_attributes_group = target.create_group('scanpath_attributes')
+        for attribute_name, attribute_value in self.scanpath_attributes.items():
+            create_hdf5_dataset(scanpath_attributes_group, attribute_name, attribute_value)
+        scanpath_attributes_group.attrs['__attributes__'] = np.string_(json.dumps(sorted(self.scanpath_attributes.keys())))
+
+        fixation_attributes_group = target.create_group('fixation_attributes')
+        for attribute_name, attribute_value in self.fixation_attributes.items():
+            fixation_attributes_group.create_dataset(attribute_name, data=attribute_value._data)
+        fixation_attributes_group.attrs['__attributes__'] = np.string_(json.dumps(sorted(self.fixation_attributes.keys())))
+
+        target.attrs['attribute_mapping'] = np.string_(json.dumps(self.attribute_mapping))
+
+
+    @classmethod
+    @hdf5_wrapper(mode='r')
+    def read_hdf5(cls, source):
+        data_type = decode_string(source.attrs['type'])
+        data_version = decode_string(source.attrs['version'])
+
+        if data_type != 'Scanpaths':
+            raise ValueError("Invalid type! Expected 'Scanpaths', got", data_type)
+
+        valid_versions = ['1.0']
+        if data_version not in valid_versions:
+            raise ValueError("Invalid version! Expected one of {}, got {}".format(', '.join(valid_versions), data_version))
+
+        lengths = source['lengths'][...]
+        xs = VariableLengthArray(source['xs'][...], lengths)
+        ys = VariableLengthArray(source['ys'][...], lengths)
+        n = source['n'][...]
+
+        scanpath_attributes = _load_attribute_dict_from_hdf5(source['scanpath_attributes'])
+
+        fixation_attributes_group = source['fixation_attributes']
+        json_attributes = fixation_attributes_group.attrs['__attributes__']
+        if not isinstance(json_attributes, str):
+            json_attributes = json_attributes.decode('utf8')
+        __attributes__ = json.loads(json_attributes)
+
+        fixation_attributes = {attribute: VariableLengthArray(fixation_attributes_group[attribute][...], lengths) for attribute in __attributes__}
+
+        return cls(
+            xs=xs,
+            ys=ys,
+            n=n,
+            lengths=lengths,
+            scanpath_attributes=scanpath_attributes,
+            fixation_attributes=fixation_attributes,
+            attribute_mapping=json.loads(decode_string(source.attrs['attribute_mapping']))
+        )
+
+    def __getitem__(self, index):
+        # TODO
+        # - integer to return single scanpath
+        # - 2d index to return single Fixation (for now via index of scanpath and index of fixation in scanpath)
+        # - 2d index array to return Fixations instance (for now via index of scanpath and index of fixation in scanpath)
+
+        if isinstance(index, tuple):
+            raise NotImplementedError("Not implemented yet")
+        elif isinstance(index, int):
+            raise NotImplementedError("Not implemented yet")
+        else:
+            return type(self)(self.xs[index], self.ys[index], self.n[index], self.lengths[index],
+                              scanpath_attributes={key: value[index] for key, value in self.scanpath_attributes.items()},
+                              fixation_attributes={key: value[index] for key, value in self.fixation_attributes.items()},
+                              attribute_mapping=self.attribute_mapping)
